lists and products eval (#9)
* implement eval with direct Expr values

* switch infer & eval to take references

no reason to take ownership of value...

this is cleaner for avoiding cloning.

* cargo fmt

* make the val/type repl printout more lisp-y

* document PrimOp trickery & fix bug

I was mistakenly not checking from initial "enclosure" of PrimOps inside
of `App`s. so I would recognize all PrimOps as being enclosed... whoops.

also, in the case of PrimOp "lifting" (packaging into a closure), no
environment is needed because no external environment is needed in the
closure, just variable bindings which are already managed.

* add lists

* implement list operations: map/fold/null

* s/TLst/TList/g

no point in contracting to 4 letters if we use TPair.

* add `TPair` type

* add pairs (product types) & accessors

* pair type pretty-print: add a space after comma

* implement `eval` for list & product types/functions

null, map, foldl, pair, fst, snd.

list also.

* add cons & nil primops

this is redundant with the list keyword.

nil doesn't work because it's a nullary function...

so now we have a weird hack where we want both `list` and `cons`.

* drop nil - it's awkward

* split to two binaries

rename binary to `polyi`. `poly` will be an interpreter for files.

* flesh out `poly` binary - interpreter from file

* switch eval & infer program functions to references

this avoid needless cloning.

* add defn processing to `polyi`

can now define functions and use them in later definitions /
expressions.

`infer`: take references to avoid cloning.

* add `nil`, revamp `list`

revamp `list` to be just a syntactic form, implemented in terms of
`cons` and `nil`. this probably reduce efficiency of interpretation,
because we had a vectorized `list` before, which packaged everything up
into a vector in one pass, with no intermediate closures.

I haven't heavily tested this but it seems to work...

* update readme to describe building/running

add examples/ directory.