fix expr normalize
we were reusing the same hashmap across parallel sub-expression
branches, which allowed for contagion / aliasing of unrelated (in terms
of lexical scoping) vars across parallel branches.

for example:

```
(let ([wat
       (lam [wat]
         (+ 1))])
  (wat 2))
```

when evaluating the `e` which is bound to `nm` (in this case
`(lam [wat] (+ 1))` being bound to `wat`), we would overwrite the
name-binder of the first `wat` (bound in the first line, in the `let`
expression). this would cause a scoping error and an `UnboundVariable`
later down in type inference of the normalized expression.