switch `panic`s to Holochain-conscious `error` macro
Holochain doesn't seem to handle `panic` well - the message is lost and
the backtrace is wasm-garbled. `hdk` provides `error` which actually
gets rendered properly. however it does not have the same polymorphic
type as `panic`. therefore I introduce `rep_lang_core::util::error`
which calls `hdk`s `error` inside of `panic`, to get the best of both
worlds. it seems to work decently well. the macro is conditional
depending on `hdk` as an optional dependencies, which means that on
non-Holochain platforms a regular `panic` will be used.