fix bug in i64 parsing
as discovered by the quickcheck properties - most excellent!

I was parsing negative numbers by registering the `-` token, then
parsing the following numeric characters as a postive i64. then, if
there was a `-` token, I would negate that postive number to make it
negative.

unfortunately this doesn't work for `minBound` of a 2s complement
representation...

i64 ranges from `-(2^63)` to `2^63 - 1`. `2^63` is not representable, so
we were getting the overflow error.